+++ client/CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Added the create a collection modal ([#655](https://github.com/project-lux/lux-frontend/issues/655)).
 - Added the delete a collection endpoint and functionality ([#660](https://github.com/project-lux/lux-frontend/issues/660)).
 - Added the Add to a Collection functionality ([#658](https://github.com/project-lux/lux-frontend/issues/658)).
+- Added My Collections entity page ([#654](https://github.com/project-lux/lux-frontend/issues/654)).
 
 ### Changed